refactor(recording): share the audio-source map and the PCM downmix
Two pieces of logic were about to gain a second copy each for the capture
handoff, so extract them first.

The cliKey -> MediaRecorder.AudioSource map decides which sources can be
captured without scrcpy. Its copy in DirectAudioRecorderSession keyed
VOICE_PERFORMANCE off "mic-voice-performance", which is not a cliKey that
exists — so that source silently fell back to scrcpy instead of using the fast
direct path. One shared map fixes it and stops the two capture paths from ever
disagreeing about what they support.

The stereo->mono downmix moves to PcmDownmix unchanged. It averages the two
call directions rather than summing them: summing overflows PCM-16 and wraps to
the opposite sign, which is an audible click on every loud moment of double-talk.
